chore: Move files to `docs/` directory
This change moves files to `docs/`

The plan is to divide the repository into two separate folders:

1. `docs/` : stores all data related to this project
2. `media/` : stores all large media. This folder is needed because GitHub does not allow files larger than 100 MB.

`media/` will NOT be stored in the GitHub repository, and its changes will NOT be tracked. The user must download these files separately from a cloud server.

The default cloud server as of 2025-08-07 is Google Drive.